refactor game board int component
